% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iocalculator.R
\name{io_classifier}
\alias{io_classifier}
\title{Sort Catch Numbers Into Species Categories}
\usage{
io_classifier(data, species = Comm.Catch.Spp.List, year = NA)
}
\arguments{
\item{data}{A data frame that includes catch numbers in dollars and TSN.}

\item{year}{A numeric variable that can be used to filter for a specific year. Defaults to NA, which returns all years.}

\item{list}{A list of lists that includes categories at the top level and TSN numbers at the second level. Defaults to Comm.Catch.Spp.List.}
}
\description{
This function uses itis_reclassify(), developed by Emily Markowitz, to create species category classifications and then sorts catch numbers into those classifications.
}
