% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iocalculator.R
\name{iocalculator}
\alias{iocalculator}
\title{Run a Fisheries Input/Output Model}
\usage{
iocalculator(base_catch, multipliers, deflator = 0.8734298, imports)
}
\arguments{
\item{base_catch}{A data frame that details catch numbers at the state-species category level for a single year, including variables fips (FIPs number, 0 for US), spec_no (a numeric variable for species category), and base_catch (raw catch numbers in dollars).}

\item{deflator}{A numeric value that adjusts jobs numbers from the current year to the year the multipliers were made; defaults to 0.8734298, which represents 2017 to 2014.}

\item{imports}{A data frame that includes imports numbers in dollars at the state-economic category level for a single year, including fips (FIPs number, 0 for US), Economic Category, and base_catch.}

\item{multiplieArs}{A data frame that includes 17 multipliers at the state-species category-economic category for a single year.}
}
\description{
This function takes in commercial fisheries catch numbers, IMPLAN multipliers, a GDP deflator, and imports numbers and outputs economic impacts.
}
